Restructure file
