+++ src/UI/Input.jsx
@@ -1,4 +1,3 @@
-// eslint-disable-next-line react/prop-types
 function Input({ type, placeholder }) {
   return (
     <input type={type} aria-label={placeholder} placeholder={placeholder} />
