+++ src/components/redux/books/BookList.jsx
@@ -1,5 +1,5 @@
 import Book from './Book';
-import Form from '../form/Form';
+import Form from '../../../UI/Form';
 
 const data = [
   {
