+++ src/components/redux/categories/Category.jsx
@@ -1,4 +1,4 @@
-import Button from '../../UI/Button';
+import Button from '../../../UI/Button';
 
 function Category() {
   return <Button type="button" title="CHECK STATUS" />;
